2026-01-12 1. AI Server -> main 업데이트
