+++ RLServer/main.py
@@ -1,23 +1,23 @@
-from stable_baselines3 import PPO
+from stable_baselines3 import SAC
 from stable_baselines3.common.vec_env import DummyVecEnv
 from stable_baselines3.common.monitor import Monitor
 from MinecraftEnv import MinecraftEnv
 
 def make_env():
-    env = MinecraftEnv((24.0, -60.0, -35.0))
+    env = MinecraftEnv((10.0, -60.0, 10.0))
     env = Monitor(env)
     return env
 
 if __name__ == "__main__":
     env = DummyVecEnv([make_env])
 
-    model = PPO.load(
-        "model/minecraft_dxdy_dz_ppo",
+    model = SAC.load(
+        "models/sac/best/best_model",
         env=env,
         device="auto"
     )
 
-    print("✅ Model loaded, inference start")
+    print("✅ SAC Model loaded, inference start")
 
     obs = env.reset()
